# routes/auth.go
package routes

import (
	"go-todo-list/config"
	"go-todo-list/models"
	"net/http"
	"time"

	"github.com/dgrijalva/jwt-go"
	"github.com/gin-gonic/gin"
)

type UserResponse struct {
    ID        uint      `json:"id"`
    Username  string    `json:"username"`
    Name      string    `json:"name"`
    CreatedAt time.Time `json:"created_at"`
    UpdatedAt time.Time `json:"updated_at"`
}

var jwtKey = []byte("your_secret_key")

// Login
// @Summary Login
// @Description Login a user and return a JWT token
// @Accept  json
// @Produce  json
// @Param login body map[string]string true "Login Input"
// @Success 200 {object} map[string]string
// @Failure 401 {object} map[string]string
// @Router /login [post]
func Login(c *gin.Context) {
    var user models.User
    var input map[string]string
    if err := c.ShouldBindJSON(&input); err != nil {
        c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
        return
    }
    username := input["username"]
    password := input["password"]

    if err := config.DB.Where("username = ?", username).First(&user).Error; err != nil {
        c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid username or password"})
        return
    }

    if !user.CheckPassword(password) {
        c.JSON(http.StatusUnauthorized, gin.H{"error": "Invalid username or password"})
        return
    }

    token := jwt.NewWithClaims(jwt.SigningMethodHS256, jwt.MapClaims{
        "username": username,
    })

    tokenString, err := token.SignedString(jwtKey)
    if err != nil {
        c.JSON(http.StatusInternalServerError, gin.H{"error": "Could not generate token"})
        return
    }

    c.JSON(http.StatusOK, gin.H{"token": tokenString})
}

// Register maneja el registro de nuevos usuarios.
// @Summary Register
// @Description Register a new user
// @Accept json
// @Produce json
// @Param user body models.User true "User Registration"
// @Success 201 {object} models.User
// @Failure 400 {object} map[string]string
// @Router /register [post]
func Register(c *gin.Context) {
    var input models.User
    if err := c.ShouldBindJSON(&input); err != nil {
        c.JSON(http.StatusBadRequest, gin.H{"error": err.Error()})
        return
    }

    // Verifica si el nombre de usuario ya existe
    var existingUser models.User
    if err := config.DB.Where("username = ?", input.Username).First(&existingUser).Error; err == nil {
        c.JSON(http.StatusBadRequest, gin.H{"error": "Username already taken"})
        return
    }

    // Cifra la contraseña
    if err := input.SetPassword(input.Password); err != nil {
        c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to encrypt password"})
        return
    }

    // Guarda el nuevo usuario en la base de datos
    if err := config.DB.Create(&input).Error; err != nil {
        c.JSON(http.StatusInternalServerError, gin.H{"error": "Failed to create user"})
        return
    }

    // Crear una respuesta sin la contraseña
    userResponse := UserResponse{
        ID:        input.ID,
        Username:  input.Username,
        Name:      input.Name,
        CreatedAt: input.CreatedAt,
        UpdatedAt: input.UpdatedAt,
    }

    c.JSON(http.StatusCreated, userResponse)
}